Open Request Page and Picking Up Requests!
The open request page displays requests that do not have a driver yet (except requests made by the user logged in, you can't deliver to yourself) and gives the ability to pick them up. The page is only accessible by drivers.  Once a request has a driver it is no longer displayed on the page, the DriverId is set to the driver's UserId, and the status of the request changes to "Accepted By Driver".