Whole Lotta Stuff That I Should Have Made Several Commits For
We have the motherload here boys. Status is fully updateable by the driver and displays correctly where appropriate. I added timed redirects to all of the confirm pages to take the user to either ReadUser, ReadDriver, or OpenRequests (although I didn't touch any of the delete stuff because I think Josh was implementing that and I didn't want to mess anything up). I added a back button to the ViewById page to take the user back to the previous page, whether it was the ReadUser or ReadDriver.  I added a further check to ReadDriver to make sure that the user is a driver.  Changed everywhere it said "Pickup Request" to "Accept Request" just to make things clearer. Made a check in UpdateDriver so that it only displays the Update Status button if the request is not delivered yet. I believe that is all.