Fixed Address Overwriting
Create method now will search for, and use an already existing address if it matches the input fields before it creates a new address. Update method no longer overwrites addresses and instead uses an already existing address or creates a new one. Both methods use user-specific addresses, meaning that an address may be in the database more than once but that is due to a different userId to associate it with their addressbook.