Image Uploading
You can now create a request with an image. You can also view the details of a request and it will display the image but that's not being done in the best way. I am currently still trying to figure out how to pass the filepath of the image from the request model back into an image in the request viewmodel. I will work on that more tomorrow.